+++ src/main.rs
@@ -3,7 +3,6 @@ use colored::Colorize;
 use env_logger::{Builder, Target};
 use fibertools_rs::cli::*;
 use fibertools_rs::subcommands;
-use fibertools_rs::subcommands::nucleosomes::add_nucleosomes_to_bam;
 use fibertools_rs::*;
 use log::LevelFilter;
 use std::time::Instant;
@@ -69,17 +68,16 @@ pub fn main() -> Result<(), Error> {
                 "Consider recompiling via cargo with: `--all-features`.",
                 "For detailed instructions see: https://fiberseq.github.io/fibertools-rs/INSTALL.html."
             );
-
             subcommands::predict_m6a::read_bam_into_fiberdata(predict_m6a_opts);
         }
         Some(Commands::ClearKinetics(clear_kinetics_opts)) => {
-            clear_kinetics(clear_kinetics_opts);
+            subcommands::clear_kinetics::clear_kinetics(clear_kinetics_opts);
         }
         Some(Commands::StripBasemods(strip_basemods_opts)) => {
             subcommands::strip_basemods::strip_base_mods(strip_basemods_opts);
         }
         Some(Commands::AddNucleosomes(nuc_opts)) => {
-            add_nucleosomes_to_bam(nuc_opts);
+            subcommands::add_nucleosomes::add_nucleosomes_to_bam(nuc_opts);
         }
         Some(Commands::Fire(fire_opts)) => {
             subcommands::fire::add_fire_to_bam(fire_opts)?;
@@ -88,10 +86,10 @@ pub fn main() -> Result<(), Error> {
             subcommands::footprint::start_finding_footprints(footprint_opts)?;
         }
         Some(Commands::Pileup(pileup_opts)) => {
-            fibertools_rs::subcommands::pileup::pileup_track(pileup_opts)?;
+            subcommands::pileup::pileup_track(pileup_opts)?;
         }
         Some(Commands::TrackDecorators(decorator_opts)) => {
-            fibertools_rs::subcommands::decorator::get_decorators_from_bam(decorator_opts)?;
+            subcommands::decorator::get_decorators_from_bam(decorator_opts)?;
         }
         Some(Commands::Completions(completion_opts)) => {
             log::info!(
@@ -101,7 +99,7 @@ pub fn main() -> Result<(), Error> {
             cli::print_completions(completion_opts.shell, &mut cli::make_cli_app());
         }
         Some(Commands::DddaToM6a(opts)) => {
-            fibertools_rs::subcommands::ddda_to_m6a::ddda_to_m6a(opts)?;
+            subcommands::ddda_to_m6a::ddda_to_m6a(opts)?;
         }
         Some(Commands::Man {}) => {
             let man = clap_mangen::Man::new(cli::make_cli_app());
