+++ src/subcommands.rs
@@ -1,5 +1,9 @@
+/// Add nucleosomes to a bam file
+pub mod add_nucleosomes;
 /// Center fiberseq information around a reference position
 pub mod center;
+/// Clear HiFi kinetics tags from a bam file
+pub mod clear_kinetics;
 pub mod ddda_to_m6a;
 /// add decorators
 pub mod decorator;
@@ -8,8 +12,6 @@ pub mod extract;
 /// add fire data
 pub mod fire;
 pub mod footprint;
-/// Add nucleosomes to a bam file
-pub mod nucleosomes;
 /// make a fire track from a bam file
 pub mod pileup;
 /// m6A prediction
