+++ src/subcommands/pileup.rs
@@ -1,6 +1,7 @@
 use crate::bamranges;
 use crate::cli::PileupOptions;
 use crate::fiber::FiberseqData;
+use crate::utils::bio_io;
 use crate::*;
 use anyhow::{self, Ok};
 use rayon::prelude::*;
