+++ tests/extract_test.rs
@@ -1,4 +1,4 @@
-use fibertools_rs::bio_io;
+use fibertools_rs::utils::bio_io;
 use fibertools_rs::utils::FiberFilters;
 use rust_htslib::bam::Read;
