+++ tests/m6a_prediction_test.rs
@@ -1,5 +1,5 @@
 use fibertools_rs;
-use fibertools_rs::bio_io;
+use fibertools_rs::utils::bio_io;
 use fibertools_rs::utils::FiberFilters;
 use rust_htslib::bam::Reader;
 use tempfile::NamedTempFile;
